Adding domains to our http server and websocket server.  This really is not the best way to go about this, but it is not super bad either.
The long story is that the ws library that I am using is emmiting errors that are not being listened to... but I don't really see any problems from ignoring them.  So in this case if the http server has an error, I think we are good to crash the server.  But the websocket may emit errors that we can ignore BUT we currently can't listen to.

This is a contrived way to listen to the wss object for errors.